Adds method for PdfValidationResult to format validation error messages
PdfValiationErrors of type UNSUPPORTED_DIMENSIONS depend on
the parent PdfValidationResult object to determine the
appropriate string formating values. This formating is
already performed in the toString()-metoden, however use
of this method is not advisable in all contexts. The
new implemented method allows for formating the
PdfValidationError using the appropriate bleed
values